Fix bug in lookForPerlFileName introduced in Komodo-PerlRemoteDebugging-7.1.3-74661
Since $result is declared in the wrong scope, "return $result" returns an undefined
global.

This affects setting a breakpoint for a file that did not go through the
postponed() hook (for examples it was loaded before the debugger, and then
Enbugger::load_source was used to hook it into the debugger) and (unverified)
getting the source code of said file.