Refactor and (hopefully) improve context_get implementation
Removed all the source code parsing heuristics, and replaced them with
(hopefully) more robust and more correct code.

It avoids returning variables with an active iterator, regardless of whether
there is a parseable each() in the source code (this might be better or worse).

It does not rely on source parsing anymore when PadWalker is not installed/disabled.