Be more compatible with Xdebug, and remove some more code
The DBGp spec is a bit vague about what property_value does for non-scalar values.

Xdebug formats the response the same way as a <property> tag,
except that there is no separate property tag, e.g.:

    <response command="property_value" transaction_id="jared"
              type="object" children="1" classname="test"
              numchildren="2">...</response>

which (to me) looks better than what perl5db.pl did until this commit.

To be honest there does not seem to be much of a point in the property_value
tag, but the spec requires it, so it stays.