Fix treatment of undef/empty strings in eval output
Before this change, both 'undef' and "" were returned with type 'string' and an
empty string as encoded value (there is not explicit way of encoding 'undef' in
DBGp).

Now they are returned with type 'undef' and 'string', so clients can make a
distinction (it might make sense to typemap 'undef' to 'null')i.