WIP - Horrible speed hack
There is a 100% overhead (in the code I'm working on at the moment) caused by
Perl calling DB::sub on each function call.

This is an attempt of reducing that overhead by injecting a C callback into
pp_entersub.

In the tradition of Devel::Declare, I'm copying and patching code from Perl
core (pp_entersub). The file core/it_could_be_worse.c contains the code for an
alternative, slower and more fragile approach that does not require copying and
patching Perl core.