The XS lsub implementation does not work for Perl earlier than 5.16
I'm not sure which commit changed behaviour, so it's entirely possible that I'm
holding call_sv() incorrectly and there is a way to make this work in XS for
older Perls.

I'm going to fallback to the Perl DB::lsub for now, until I have more time to
look into this.