t/152_property_set.t - Handle longer error message on 5.37.9 and later
Errors are now consistently handled in eval, regardless of count, or
type of error (semantic or syntax, etc). In older perls a single syntax
error would not trigger $SIG{__DIE__} for instance, nor would it exit
the eval the same way as 10+ errors would. In 5.37.9 all errors are
handled the same, and will trigger $SIG{__DIE__}.